pyproject.toml
[tool.poetry]
name = "complexpy"
version = "0.1.0"
description = "This package allows to apply several measures of emergence and complexity to either empirical or simulated time-series data, and provide guidance for comparisons among and conclusions about different measures."
authors = ["Nadine Spychala"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "complexpy", from = "src"}
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
matlabengine = "^23.2.1"
pandas = "^2.1.3"
numpy = "^1.26.2"
scipy = "^1.11.4"
pytest = "^7.4.3"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
